Update commit of what I have been working on today. I have been working on creating Habitats and have them interact     the animals. I have used Aggregation ( http://en.wikipedia.org/wiki/Class_diagram#Aggregation ) class diagram idea thingy. Habitat <>---- Animal
it has been quite tough working through this.

ArrayLists are new for me and worked with those.

The code seems one big mess right now, and I barely have any clear overview of the project at the moment.

Right now I am going to try and structure the files in different folder/directories/packages

This commit is also a backup :)